3.5.3-SNAPSHOT - 02/01/2018
Removed the hibernate-search-analyzers dependency from AppFuse. This
artifact has been deprecated since Hibernate 4, and it includes an extra
version of Lucene that causes problems with RSCMGR deployment. Updated
Hibernate to use the hibernate-search-orm artifact, which is the updated
package for hibernate search. Removed dependencies for hibernate core,
as this is now included by the hibernate-search-orm.